+++ src/visible.h
@@ -4,14 +4,14 @@
 #include "ray.h"
 
 struct collision {
-	float dist;
+	double dist;
 	vec3 point;
 	vec3 normal;
 };
 
 class visible {
 public:
-	virtual bool intersects(const ray& r, float dist_min, float dist_max, collision& data) const = 0;
+	virtual bool intersects(const ray& r, double dist_min, double dist_max, collision& data) const = 0;
 };
 
 #endif
